third_party/WebKit/LayoutTests/fast/forms/calendar-picker/calendar-picker-should-not-change-datetimelocal-time.html

<!DOCTYPE html>
<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
<script src="../../forms/resources/picker-common.js"></script>
<script src="../resources/common.js"></script>
<script src="resources/calendar-picker-common.js"></script>
</head>
<body>
<script>
description('Test for crbug.com/271451. The calendar picker should not change the time part of datetime-local field.');

var input = document.createElement('input');
input.type = 'datetime-local';
getElementByPseudoId(internals.youngestShadowRoot(input), '-webkit-datetime-edit').setAttribute('pattern', 'yyyy-MM-dd HH:mm');
document.body.appendChild(input);

input.value = '1999-07-31T12:59';

shouldBeEqualToString('getUserAgentShadowTextContent(input)', '1999-07-31 12:59');
openPicker(input, function() {
    shouldBeEqualToString('getUserAgentShadowTextContent(input)', '1999-07-31 12:59');

    eventSender.keyDown('rightArrow');
    shouldBeEqualToString('highlightedValue()', '1999-08-01');

    eventSender.keyDown('\n');
    shouldBeEqualToString('selectedValue()', '1999-08-01');

    shouldBeEqualToString('getUserAgentShadowTextContent(input)', '1999-08-01 12:59');
    finishJSTest();
});

</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
